test/mempool: decrease required memory
To pass the mempool test, hundreds of megabytes are needed (the real
amount depends on the platform).

In this functional test, the number of elements in mempool does not need
to depend on the number of lcores, as we only run on the master
lcore. Furthermore, 2KB for each object is quite large, reduce it by 8.

On my platform, it divides by ~100 the amount of required memory to pass
the test.

Signed-off-by: Olivier Matz <olivier.matz@6wind.com>