mbuf: fix reset on mbuf free
m->nb_seg must be reset on mbuf free whatever the value of m->next,
because it can happen that m->nb_seg is != 1. For instance in this
case:

  m1 = rte_pktmbuf_alloc(mp);
  rte_pktmbuf_append(m1, 500);
  m2 = rte_pktmbuf_alloc(mp);
  rte_pktmbuf_append(m2, 500);
  concat(m1, m2);
  m0 = rte_pktmbuf_alloc(mp);
  rte_pktmbuf_append(m0, 500);
  concat(m0, m1);
  new_m = split(m0, offset=1000)

If concat() and split() do not reset nb_seg in the initial m1 segment,
after this code, we have 2 mbuf chain:
  - m0 with 2 segments, the last one has next=NULL but nb_seg=2
  - new_m with 1 segment

Freeing m0 will not restore nb_seg=1 in the second segment.

Fixes: 8f094a9ac5d7 ("mbuf: set mbuf fields while in pool")
Cc: stable@dpdk.org

Signed-off-by: Olivier Matz <olivier.matz@6wind.com>