mempool: fix mempool drivers usage in secondary
When the mempool drivers in the primary applications are not loaded in
the same order than in the secondary application, the behavior is
undefined because the mempool->ops_index corresponds to different
drivers.

This patch adds a table in shared memory, which associates an index to a
mempool driver. This table is filled by the primary process, and
accessed by secondary processes when mempool ops are registered.

Unfortunatly, rte_mempool_register_ops() cannot be called too early (in
a constructor), because it needs to access to access to the shared
memory. A previous tentative to solve this issue was introducing a new
init level. This patch instead uses a vdev for each mempool driver,
and take advantage of the probing method which is called by EAL at the
right moment.

The RTE_MEMPOOL_REGISTER_OPS() macro is moved in a new file, because
it includes some files that are part of "driver_sdk_headers".

There is no ABI change from the application perspective, but after this
change, the mempool drivers have to be recompiled.

Link: http://inbox.dpdk.org/dev/1586787714-13890-1-git-send-email-xiangxia.m.yue@gmail.com/
Signed-off-by: Olivier Matz <olivier.matz@6wind.com>