bus/vdev: new macro to instanciate a vdev at startup
The next commit will rework the way mempool drivers register their
ops by using a vdev which is always present.

Introduce a macro to create a virtual device at startup: when the
vdev bus is probed, this device is added automatically and is directly
usable, without requiring the --vdev command line argument.

Signed-off-by: Olivier Matz <olivier.matz@6wind.com>